Added: Exception to the application

+++ java/com/example/employeepayrollapplication/services/EmployeePayRollService.java
@@ -1,6 +1,7 @@
 package com.example.employeepayrollapplication.services;
 
 import com.example.employeepayrollapplication.dto.EmployeePayrollDTO;
+import com.example.employeepayrollapplication.exception.EmployeeNotFoundException;
 import com.example.employeepayrollapplication.model.EmployeePayrollData;
 import com.example.employeepayrollapplication.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,8 +24,13 @@ public class EmployeePayRollService implements IEmployeePayrollService{
     }
 
     @Override
-    public Optional<EmployeePayrollData> getEmployeePayRollDataById(int empId) {
-        return employeeRepository.findById(empId);
+    public EmployeePayrollData getEmployeePayRollDataById(int empId) throws EmployeeNotFoundException {
+        EmployeePayrollData employeePayrollData =  employeeRepository.findByEmployeeId(empId);
+        if(employeePayrollData != null) {
+            return employeePayrollData;
+        } else {
+            throw new EmployeeNotFoundException("Employee not found with id: " + empId);
+        }
     }
 
     @Override
@@ -36,12 +42,17 @@ public class EmployeePayRollService implements IEmployeePayrollService{
     }
 
     @Override
-    public void deleteEmployeePayRollDataById(int empId) {
-        employeeRepository.deleteById(empId);
+    public void deleteEmployeePayRollDataById(int empId) throws EmployeeNotFoundException {
+        EmployeePayrollData employeePayrollData = employeeRepository.findByEmployeeId(empId);
+        if(employeePayrollData != null) {
+            employeeRepository.deleteById(empId);
+        } else {
+            throw new EmployeeNotFoundException("Employee not found with id: " + empId);
+        }
     }
 
     @Override
-    public EmployeePayrollData updateEmployeePayRollData(int empId, EmployeePayrollDTO employeePayrollDTO) {
+    public EmployeePayrollData updateEmployeePayRollData(int empId, EmployeePayrollDTO employeePayrollDTO) throws EmployeeNotFoundException {
         List<EmployeePayrollData> employeePayrollDataList = this.getEmployeePayRollData();
         for (EmployeePayrollData empData : employeePayrollDataList) {
                 if(empData.getEmployeeId() == empId) {
@@ -49,6 +60,8 @@ public class EmployeePayRollService implements IEmployeePayrollService{
                     empData.setSalary(employeePayrollDTO.salary);
                     employeeRepository.save(empData);
                     return empData;
+                } else {
+                    throw new EmployeeNotFoundException("Employee not found with id: " + empId);
                 }
         }
         return null;
